Move the identical from ActiveForm::Base and ActiveForm::Form into a module name ActiveForm::FormHelpers.
