+++ lib/active_form/base.rb
@@ -1,6 +1,7 @@
 module ActiveForm
   class Base
     include ActiveModel::Model
+    include FormHelpers
     extend ActiveModel::Callbacks
 
     define_model_callbacks :save, only: [:after]
@@ -14,17 +15,7 @@ module ActiveForm
       @forms = []
       populate_forms
     end
-
-    def submit(params)
-      params.each do |key, value|
-        if nested_params?(value)
-          fill_association_with_attributes(key, value)
-        else
-          send("#{key}=", value)
-        end
-      end
-    end
-
+    
     def get_model(assoc_name)
       form = find_form_by_assoc_name(assoc_name)
       form.get_model(assoc_name)
@@ -40,16 +31,6 @@ module ActiveForm
       else
         false
       end
-    end
-
-    def valid?
-      super
-      model.valid?
-
-      collect_errors_from(model)
-      aggregate_form_errors
-
-      errors.empty?
     end
 
     class << self
@@ -79,26 +60,17 @@ module ActiveForm
       alias_method :attribute, :attributes
 
       def association(name, options={}, &block)
+        forms << FormDefinition.new(name, block, options)
         macro = main_class.reflect_on_association(name).macro
 
         case macro
         when :has_one, :belongs_to
-          declare_form(name, &block)
+          class_eval "def #{name}; @#{name}; end"
         when :has_many
-          declare_form_collection(name, options, &block)
+          class_eval "def #{name}; @#{name}.models; end"
         end
 
-        define_method("#{name}_attributes=") {}
-      end
-
-      def declare_form_collection(name, options={}, &block)
-        forms << FormDefinition.new(name, block, options)
-        class_eval("def #{name}; @#{name}.models; end")
-      end
-
-      def declare_form(name, &block)
-        forms << FormDefinition.new(name, block)
-        attr_reader name
+        class_eval "def #{name}_attributes=; end"
       end
 
       def forms
@@ -124,45 +96,5 @@ module ActiveForm
       end
     end
 
-    def nested_params?(value)
-      value.is_a?(Hash)
-    end
-
-    ATTRIBUTES_KEY_REGEXP = /^(.+)_attributes$/
-
-    def find_association_name_in(key)
-      ATTRIBUTES_KEY_REGEXP.match(key)[1]
-    end
-
-    def fill_association_with_attributes(association, attributes)
-      assoc_name = find_association_name_in(association).to_sym
-      form = find_form_by_assoc_name(assoc_name)
-
-      form.submit(attributes)
-    end
-
-    def find_form_by_assoc_name(assoc_name)
-      forms.select { |form| form.represents?(assoc_name) }.first
-    end
-
-    def aggregate_form_errors
-      forms.each do |form|
-        form.valid?
-        collect_errors_from(form)
-      end
-    end
-
-    def collect_errors_from(validatable_object)
-      validatable_object.errors.each do |attribute, error|
-        key = if validatable_object.respond_to?(:association_name)
-          "#{validatable_object.association_name}.#{attribute}"
-        else
-          attribute
-        end
-
-        errors.add(key, error)
-      end
-    end
   end
-
 end
